Add uninstalling apps to installation proxy service
