+++ Netimobiledevice/InstallationProxy/AppInstallException.cs
@@ -1,0 +1,11 @@
+﻿using Netimobiledevice.Exceptions;
+
+namespace Netimobiledevice.InstallationProxy
+{
+    public class AppInstallException : NetimobiledeviceException
+    {
+        public AppInstallException() : base() { }
+
+        public AppInstallException(string message) : base(message) { }
+    }
+}
